feat: APIservice modificado, Componentes pagina creados para desaarrollar
