+++ app_mindker/src/components/Register/Register.jsx
@@ -19,7 +19,7 @@ import { AiFillContacts } from 'react-icons/ai';
 import { useNavigate } from 'react-router-dom';
 
 import GlobalContext from '../../context/GlobalContext';
-import { loginUser, RegisterUser } from '../../services/postsFunctionsApiUser.js';
+import { loginUser, RegisterUser } from '../../services/APIservice.js';
 import AgnosticButton from '../AgnosticButton/AgnosticButton';
 
 const Register = () => {
