+++ app_mindker/src/pages/dashboard/Dashboard.jsx
@@ -1,18 +1,18 @@
-import { useContext, useEffect, useRef, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 
 import DecksSuperContainer from '../../components/DecksContainer/DecksSuperContainer';
 import NavUserDashboard from '../../components/NavUserDashboard/NavUserDashboard';
 import GlobalContext from '../../context/GlobalContext';
 import DashboardLayout from '../../layouts/DasboardLayout/DashboardLayout';
-import { getAgnostic } from '../../services/getServices';
+import { getAgnostic } from '../../services/APIservice';
 
 const Dashboard = () => {
-  const { user, dashboardContent, switcher, param, setParam, paramReforce } =
+  const { user, dashboardContent, switcher, param, setParam, local, paramReforce } =
     useContext(GlobalContext);
   // eslint-disable-next-line no-unsafe-optional-chaining
   const allUserDecks = [...user?.downloadedDecks, ...user?.createdDecks];
-  const ref = useRef(allUserDecks);
-  const [arrayDecks, setArrayDecks] = useState(ref);
+  //const ref = useRef(allUserDecks);
+  const [arrayDecks, setArrayDecks] = useState(allUserDecks);
   const [textDecks, setTextDecks] = useState('My Decks');
 
   useEffect(() => {
@@ -24,7 +24,7 @@ const Dashboard = () => {
       setTextDecks(param);
     dashboardContent &&
       getAgnostic(dashboardContent, param).then((res) => setArrayDecks(res.info.data));
-  }, [switcher]);
+  }, [switcher, local]);
 
   return (
     <DashboardLayout direction="row">
