+++ app_mindker/src/pages/detailDeck/DetailDeck.jsx
@@ -1,0 +1,5 @@
+const DetailDeck = () => {
+  return <div>DetailDeck</div>;
+};
+
+export default DetailDeck;
