--- app_mindker/src/pages/updateProfile/UpdateProfile.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UpdateProfile = () => {
-  return <div>UpdateProfile</div>;
-};
-
-export default UpdateProfile;
